refactor(puzzles): extract live fallback builders
